# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service for collecting metrics from agents",
        "title": "MetricService API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/value/{metricType}/{metricName}": {
            "get": {
                "description": "Get metric information by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get metric information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "gauge or counter",
                        "name": "metricType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name of metric",
                        "name": "metricName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Metrics"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Metrics": {
            "type": "object",
            "properties": {
                "delta": {
                    "description": "значение метрики в случае передачи counter",
                    "type": "integer"
                },
                "hash": {
                    "description": "значение хеш-функции",
                    "type": "string"
                },
                "id": {
                    "description": "имя метрики",
                    "type": "string"
                },
                "type": {
                    "description": "параметр, принимающий значение gauge или counter",
                    "type": "string"
                },
                "value": {
                    "description": "значение метрики в случае передачи gauge",
                    "type": "number"
                }
            }
        }
    }
}